Split backend graphql schema layer by domain (#259)
- schemas/converters.py: all DTO builders moved out of queries.py, public names
- per-domain modules mirroring app/repositories/: project, imports, po,
  warehouse, shipping, shop_assembly, stock, vendor, relay, user,
  notification, dashboard, admin
- queries.py / mutations.py reduced to thin composition via inheritance
- inline-sqlalchemy resolvers normalized to repository functions
  (list_projects_with_opening_counts, get_open_pos, reload_po, find_vendor,
  find_warehouse, get_packing_slip, get_shipment_return,
  get_openings_with_items, get_request_with_openings, list_excluded_items)
- schema verified field-identical: 62 query + 53 mutation fields
