+++ backend/app/schemas/shipping.py
@@ -1,0 +1,125 @@
+"""Shipping queries + mutations: ship-ready items, packing slips, returns."""
+
+import uuid
+
+import strawberry
+
+from app.database import SessionLocal
+from app.repositories import shipping_repository
+
+from .converters import opening_item_to_type, packing_slip_to_type, shipment_return_to_type
+from .inputs import ConfirmShipmentInput, CreateShipmentReturnInput
+from .types import (
+    PackingSlip,
+    ReturnableLine,
+    ShipmentReturn,
+    ShipReadyItems,
+    ShipReadyLooseItem,
+)
+
+
+@strawberry.type
+class ShippingQueries:
+    @strawberry.field
+    def ship_ready_items(self, project_id: strawberry.ID | None = None) -> ShipReadyItems:
+        with SessionLocal() as session:
+            data = shipping_repository.get_ship_ready_items(session, uuid.UUID(str(project_id)) if project_id else None)
+            return ShipReadyItems(
+                opening_items=[opening_item_to_type(oi) for oi in data["opening_items"]],
+                loose_items=[
+                    ShipReadyLooseItem(
+                        opening_number=li["opening_number"],
+                        hardware_category=li["hardware_category"],
+                        product_code=li["product_code"],
+                        available_quantity=li["available_quantity"],
+                    )
+                    for li in data["loose_items"]
+                ],
+            )
+
+    @strawberry.field
+    def packing_slips(self, project_id: strawberry.ID | None = None) -> list[PackingSlip]:
+        with SessionLocal() as session:
+            slips = shipping_repository.list_packing_slips(session, uuid.UUID(str(project_id)) if project_id else None)
+            return [packing_slip_to_type(ps) for ps in slips]
+
+    @strawberry.field
+    def returnable_lines(self, packing_slip_id: strawberry.ID) -> list[ReturnableLine]:
+        with SessionLocal() as session:
+            lines = shipping_repository.get_returnable_lines(session, uuid.UUID(str(packing_slip_id)))
+            return [
+                ReturnableLine(
+                    packing_slip_item_id=strawberry.ID(str(line["packing_slip_item_id"])),
+                    opening_number=line["opening_number"],
+                    product_code=line["product_code"],
+                    hardware_category=line["hardware_category"],
+                    shipped_quantity=line["shipped_quantity"],
+                    returned_quantity=line["returned_quantity"],
+                    returnable_quantity=line["returnable_quantity"],
+                )
+                for line in lines
+            ]
+
+
+@strawberry.type
+class ShippingMutations:
+    @strawberry.mutation
+    def confirm_shipment(self, input: ConfirmShipmentInput) -> PackingSlip:
+        from app.models.enums import PullRequestItemType
+
+        project_id = uuid.UUID(str(input.project_id))
+        items_data = []
+        for item in input.items:
+            d = {
+                "item_type": PullRequestItemType(item.item_type.value),
+                "quantity": item.quantity,
+            }
+            if item.opening_item_id is not None:
+                d["opening_item_id"] = uuid.UUID(str(item.opening_item_id))
+            if item.opening_number is not None:
+                d["opening_number"] = item.opening_number
+            if item.product_code is not None:
+                d["product_code"] = item.product_code
+            if item.hardware_category is not None:
+                d["hardware_category"] = item.hardware_category
+            items_data.append(d)
+
+        with SessionLocal() as session:
+            ps = shipping_repository.confirm_shipment(
+                session,
+                project_id,
+                input.packing_slip_number,
+                input.shipped_by,
+                items_data,
+            )
+            session.commit()
+            refreshed = shipping_repository.get_packing_slip(session, ps.id)
+            return packing_slip_to_type(refreshed)
+
+    @strawberry.mutation
+    def create_shipment_return(self, input: CreateShipmentReturnInput) -> ShipmentReturn:
+        from app.models.enums import ReturnDisposition
+
+        items_data = [
+            {
+                "packing_slip_item_id": uuid.UUID(str(it.packing_slip_item_id)),
+                "quantity": it.quantity,
+                "disposition": ReturnDisposition(it.disposition.value),
+                "rma_reference": it.rma_reference,
+                "reason_text": it.reason_text,
+            }
+            for it in input.items
+        ]
+
+        with SessionLocal() as session:
+            sr = shipping_repository.create_shipment_return(
+                session,
+                packing_slip_id=uuid.UUID(str(input.packing_slip_id)),
+                warehouse_id=uuid.UUID(str(input.warehouse_id)),
+                returned_by=input.returned_by,
+                reference=input.reference,
+                items=items_data,
+            )
+            session.commit()
+            refreshed = shipping_repository.get_shipment_return(session, sr.id)
+            return shipment_return_to_type(refreshed)
